[REFACTOR] 패키징 정리
