+++ app/src/main/java/com/example/androidseminar/fragments/FollowerFragment.kt
@@ -14,15 +14,17 @@ import com.example.androidseminar.R
 import com.example.androidseminar.adapter.FollowerRecyclerViewAdapter
 import com.example.androidseminar.api.GithubCreator
 import com.example.androidseminar.data.Info
-import com.example.androidseminar.data.ResponseGithubBioData
 import com.example.androidseminar.data.ResponseGithubFollowerData
 import com.example.androidseminar.databinding.FragmentFollowerBinding
 import com.example.androidseminar.util.*
+import com.example.androidseminar.util.function.HorizontalItemDecoration
+import com.example.androidseminar.util.function.ItemDragListener
+import com.example.androidseminar.util.function.ItemTouchHelperCallback
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class FollowerFragment :ItemDragListener, BaseFragment<FragmentFollowerBinding>() {
+class FollowerFragment : ItemDragListener, BaseFragment<FragmentFollowerBinding>() {
 
     private lateinit var adapter: FollowerRecyclerViewAdapter
     private lateinit var itemTouchHelper: ItemTouchHelper
