+++ app/src/main/java/com/example/androidseminar/util/function/ItemTouchHelperCallback.kt
@@ -1,8 +1,7 @@
-package com.example.androidseminar.util
+package com.example.androidseminar.util.function
 
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
-import com.example.androidseminar.util.ItemActionListener
 
 class ItemTouchHelperCallback(val listener: ItemActionListener) : ItemTouchHelper.Callback() {
     override fun getMovementFlags(
