+++ app/src/main/java/com/example/androidseminar/util/room/UserDao.kt
@@ -1,4 +1,4 @@
-package com.example.androidseminar.room
+package com.example.androidseminar.util.room
 
 import androidx.room.Delete
 import androidx.room.Insert
