+++ app/src/main/java/com/example/androidseminar/util/room/UserDatabase.kt
@@ -1,4 +1,4 @@
-package com.example.androidseminar.room
+package com.example.androidseminar.util.room
 
 import android.content.Context
 import androidx.room.Database
